requirements.txt
groq>=1.0.0
openai>=1.13.0
anthropic>=0.30.0
google-generativeai>=0.3.0
requests>=2.31.0
python-dotenv>=1.0.0
httpx>=0.25.0
